Fix&Refactor: error of NSPIN=2 and refactor DensityMatrix class
