+++ source/Makefile.Objects
@@ -230,6 +230,7 @@ OBJS_ELECSTAT_LCAO=elecstate_lcao.o\
       elecstate_lcao_tddft.o\
       elecstate_lcao_cal_tau.o\
       density_matrix.o\
+      density_matrix_io.o\
       cal_dm_psi.o\
 
 OBJS_ESOLVER=esolver.o\
