+++ source/module_elecstate/CMakeLists.txt
@@ -35,6 +35,7 @@ if(ENABLE_LCAO)
       elecstate_lcao_cal_tau.cpp
       potentials/H_TDDFT_pw.cpp
       module_dm/density_matrix.cpp
+      module_dm/density_matrix_io.cpp
       module_dm/cal_dm_psi.cpp
   )
 endif()
