+++ source/module_elecstate/module_dm/test/CMakeLists.txt
@@ -11,7 +11,7 @@ endif()
 AddTest(
   TARGET dm_io_test_serial
   LIBS parameter ${math_libs} base device cell_info 
-  SOURCES test_dm_io.cpp ../density_matrix.cpp
+  SOURCES test_dm_io.cpp ../density_matrix.cpp ../density_matrix_io.cpp
   ${ABACUS_SOURCE_DIR}/module_hamilt_lcao/module_hcontainer/base_matrix.cpp
   ${ABACUS_SOURCE_DIR}/module_hamilt_lcao/module_hcontainer/hcontainer.cpp
   ${ABACUS_SOURCE_DIR}/module_hamilt_lcao/module_hcontainer/atom_pair.cpp
@@ -23,7 +23,7 @@ AddTest(
 AddTest(
   TARGET dm_constructor_test
   LIBS parameter  ${math_libs} base device
-  SOURCES test_dm_constructor.cpp ../density_matrix.cpp tmp_mocks.cpp
+  SOURCES test_dm_constructor.cpp ../density_matrix.cpp ../density_matrix_io.cpp tmp_mocks.cpp
   ${ABACUS_SOURCE_DIR}/module_hamilt_lcao/module_hcontainer/base_matrix.cpp
   ${ABACUS_SOURCE_DIR}/module_hamilt_lcao/module_hcontainer/hcontainer.cpp
   ${ABACUS_SOURCE_DIR}/module_hamilt_lcao/module_hcontainer/atom_pair.cpp
@@ -34,7 +34,7 @@ AddTest(
 AddTest(
   TARGET dm_init_test
   LIBS parameter  ${math_libs} base device
-  SOURCES test_dm_R_init.cpp ../density_matrix.cpp tmp_mocks.cpp
+  SOURCES test_dm_R_init.cpp ../density_matrix.cpp ../density_matrix_io.cpp tmp_mocks.cpp
   ${ABACUS_SOURCE_DIR}/module_hamilt_lcao/module_hcontainer/base_matrix.cpp
   ${ABACUS_SOURCE_DIR}/module_hamilt_lcao/module_hcontainer/hcontainer.cpp
   ${ABACUS_SOURCE_DIR}/module_hamilt_lcao/module_hcontainer/atom_pair.cpp
@@ -45,7 +45,7 @@ AddTest(
 AddTest(
   TARGET dm_cal_DMR_test
   LIBS parameter  ${math_libs} base device
-  SOURCES test_cal_dm_R.cpp ../density_matrix.cpp tmp_mocks.cpp
+  SOURCES test_cal_dm_R.cpp ../density_matrix.cpp ../density_matrix_io.cpp tmp_mocks.cpp
   ${ABACUS_SOURCE_DIR}/module_hamilt_lcao/module_hcontainer/base_matrix.cpp
   ${ABACUS_SOURCE_DIR}/module_hamilt_lcao/module_hcontainer/hcontainer.cpp
   ${ABACUS_SOURCE_DIR}/module_hamilt_lcao/module_hcontainer/atom_pair.cpp
