+++ source/module_hamilt_lcao/hamilt_lcaodft/edm.cpp
@@ -99,6 +99,5 @@ elecstate::DensityMatrix<std::complex<double>, double> Force_LCAO<std::complex<d
     // cal_dm_2d
     edm.init_DMR(ra, &ucell);
     edm.cal_DMR();
-    // edm.sum_DMR_spin();
     return edm;
 }
